+++ examples/02 - hello_world_chrome/hello.py
@@ -2,8 +2,6 @@ import os
 import sys
 
 sys.path.insert(0, os.path.join(os.path.dirname(os.path.abspath(__file__)), '..'))
-from common import WINDOW_SIZE
-
 import glue
 
 # Serve ui/ (default)
@@ -17,4 +15,4 @@ say_hello_py('Python World!')
 glue.say_hello_js('Python World!')   # Call a Javascript function
 
 # Force Google Chrome/Chromium (app mode still on by default)
-glue.start('hello.html', mode='chrome', size=WINDOW_SIZE)
+glue.start('hello.html', mode='chrome')