+++ examples/07 - jinja_templates/hello.py
@@ -3,8 +3,6 @@ import sys
 import random
 
 sys.path.insert(0, os.path.join(os.path.dirname(os.path.abspath(__file__)), '..'))
-from common import WINDOW_SIZE
-
 import glue
 
 glue.init()                     # Serve the ui/ folder
@@ -20,4 +18,4 @@ def say_hello_py(x):
 say_hello_py('Python World!')
 glue.say_hello_js('Python World!')   # Call a Javascript function
 
-glue.start('templates/hello.html', size=WINDOW_SIZE, jinja_templates='templates')
+glue.start('templates/hello.html', jinja_templates='templates')