+++ examples/08 - createreactapp/glue_CRA.py
@@ -62,7 +62,6 @@ def start_glue(develop):
         mode=mode,
         host='localhost',
         port=8080,
-        size=(1280, 720),
     )